Renaming dataclasses folder.
