+++ src/adler/__init__.py
@@ -1,4 +1,4 @@
-from . import dataclasses
+from . import objectdata
 from . import science
 from . import utilities
 from . import lasair
