+++ src/adler/objectdata/AdlerData.py
