+++ src/adler/objectdata/MPCORB.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from adler.dataclasses.dataclass_utilities import get_from_table, get_from_dictionary
+from adler.objectdata.objectdata_utilities import get_from_table, get_from_dictionary
 
 MPCORB_KEYS = {
     "mpcDesignation": str,
