+++ src/adler/objectdata/__init__.py
@@ -1,6 +1,6 @@
 from . import AdlerData
 from . import Observations
-from . import dataclass_utilities
+from . import objectdata_utilities
 from . import MPCORB
 from . import AdlerPlanetoid
 from . import SSObject
