+++ src/adler/objectdata/objectdata_utilities.py
